change fileUploadState impl
